[IDLE-107] rename: file 관련 예외 에러 코드 및 패키지명 수정
  - 패키지명) file -> io
  - 에러 코드) F-01 -> IO-01
  - follow 관련 예외와 에러 코드가 겹칠 것 같아서 수정
