+++ user-service/src/main/java/kr/mybrary/userservice/user/domain/exception/io/EmptyFileException.java
@@ -1,11 +1,11 @@
-package kr.mybrary.userservice.user.domain.exception.file;
+package kr.mybrary.userservice.user.domain.exception.io;
 
 import kr.mybrary.userservice.global.exception.ApplicationException;
 
 public class EmptyFileException extends ApplicationException {
 
     private final static int STATUS = 400;
-    private final static String ERROR_CODE = "F-01";
+    private final static String ERROR_CODE = "IO-01";
     private final static String ERROR_MESSAGE = "파일이 비어있습니다. 파일을 선택해주세요.";
 
     public EmptyFileException() {
