+++ user-service/src/main/java/kr/mybrary/userservice/user/domain/exception/io/FileInputStreamException.java
@@ -1,11 +1,11 @@
-package kr.mybrary.userservice.user.domain.exception.file;
+package kr.mybrary.userservice.user.domain.exception.io;
 
 import kr.mybrary.userservice.global.exception.ApplicationException;
 
 public class FileInputStreamException extends ApplicationException {
 
     private final static int STATUS = 500;
-    private final static String ERROR_CODE = "F-02";
+    private final static String ERROR_CODE = "IO-02";
     private final static String ERROR_MESSAGE = "파일을 읽어오는데 실패했습니다.";
 
     public FileInputStreamException() {
